Move ical property types to separate file
