+++ src/utils/escapeText.ts
@@ -1,4 +1,4 @@
-const escapeText = text => {
+const escapeText = (text: string) => {
   if (!text) return ''
   return text
     .replaceAll(',', '\\,')
