+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as formatDate } from './formatDate'
 export { default as escapeText } from './escapeText'
+export { default as takeOr } from './takeOr'
